Time: 40 ms (5.49%) | Memory: 11.1 MB (76.49%) - LeetSync

+++ 1263-number-of-dice-rolls-with-target-sum/number-of-dice-rolls-with-target-sum.cpp
@@ -2,7 +2,7 @@ class Solution {
 public:
     int numRollsToTarget(int n, int k, int target) {
         int a = 1e9+7;
-        vector<vector<long long >>dp(n+1,vector<long long>(target+1,0));
+        vector<vector<int>>dp(n+1,vector<int>(target+1,0));
         dp[0][0] = 1;
         for(int i = 1;i<=n;i++){
             for(int j = 1;j<=target;j++){
